requirements.txt
streamlit
pandas
psycopg2-binary
folium
streamlit-folium